move validation to telemetry config
